Move test machinery to different submodule
